decoder -> decompress-frame
